fix(compile): resolve 9 rustc errors - LoadedDll passthrough, Option<Arc> assign, oneshot flatten, borrow split in handle_host_hello, const ip_v4, u32-vip cast, set_tos_v4

+++ src-tauri/src/lobby/room.rs
@@ -40,7 +40,7 @@ impl Room {
         while self.next_host_byte <= 254 {
             let b = self.next_host_byte;
             self.next_host_byte += 1;
-            let vip_u32 = ip_v4(10, 88, 0, b);
+            let vip_u32 = ip_v4(10, 88, 0, b as u8);
             let vip = Ipv4Addr::from(vip_u32).to_string();
             let taken = self.peers.values().any(|p| p.vip == vip) || self.self_peer.vip == vip;
             if !taken {
